# norm for Vec