Master:
```jl
julia> @benchmark norm($a)
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     999
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     9.00 ns (0.00% GC)
  median time:      10.00 ns (0.00% GC)
  mean time:        10.37 ns (0.00% GC)
  maximum time:     57.00 ns (0.00% GC)
```
PR:
```jl
julia> @benchmark norm($a)
BenchmarkTools.Trial: 
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     2.00 ns (0.00% GC)
  median time:      2.00 ns (0.00% GC)
  mean time:        2.07 ns (0.00% GC)
  maximum time:     41.00 ns (0.00% GC)
```

Gotta cut 'em nanoseconds. 500% faster Kreygasm.