@fredrikekre 2016-10-31
Master:
```jl
julia> @code_native norm(a)
	.text
Filename: generic.jl
	pushq	%rbp
	movq	%rsp, %rbp
Source line: 201
	movabsq	$vecnorm, %rax
	movl	$2, %esi
	callq	*%rax
	popq	%rbp
	retq
	nopw	(%rax,%rax)
```

PR:
```jl
julia> @code_native norm(a)
	.text
Filename: tensor_ops.jl
	pushq	%rbp
	movq	%rsp, %rbp
	pushq	%r15
	pushq	%r14
	pushq	%r12
	pushq	%rbx
	subq	$80, %rsp
	movabsq	$139950764596344, %r12  # imm = 0x7F48D39C1C78
	movq	%fs:0, %r15
	addq	$-2672, %r15            # imm = 0xFFFFFFFFFFFFF590
	movq	$0, -88(%rbp)
	xorps	%xmm0, %xmm0
	movups	%xmm0, -56(%rbp)
	movups	%xmm0, -72(%rbp)
	movq	$0, -40(%rbp)
	movq	$14, -104(%rbp)
	movq	(%r15), %rax
	movq	%rax, -96(%rbp)
	leaq	-104(%rbp), %rax
	movq	%rax, (%r15)
	movq	$0, -80(%rbp)
Source line: 46
	movabsq	$139950488862784, %rbx  # imm = 0x7F48C32CC040
	movq	(%rbx), %rax
	testq	%rax, %rax
	jne	L142
	leaq	98036888(%r12), %rdi
	movabsq	$jl_get_binding_or_error, %rax
	movq	%r12, %rsi
	callq	*%rax
	movq	%rax, (%rbx)
L142:
	movq	8(%rax), %rbx
	testq	%rbx, %rbx
	je	L306
	movq	%rbx, -88(%rbp)
	movabsq	$139950488862792, %r14  # imm = 0x7F48C32CC048
	movq	(%r14), %rax
	testq	%rax, %rax
	jne	L203
	leaq	98036888(%r12), %rdi
	movabsq	$jl_get_binding_or_error, %rax
	movq	%r12, %rsi
	callq	*%rax
	movq	%rax, (%r14)
L203:
	movq	8(%rax), %rax
	testq	%rax, %rax
	je	L321
	leaq	-72(%rbp), %rdi
	leaq	-48(%rbp), %r14
	movq	%rax, -80(%rbp)
	leaq	44188872(%r12), %rcx
	movq	%rcx, -72(%rbp)
	movq	%rbx, -64(%rbp)
	movq	%rax, -56(%rbp)
	movabsq	$jl_apply_generic, %rbx
	movl	$3, %esi
	callq	*%rbx
	movq	%rax, -40(%rbp)
	addq	$44891336, %r12         # imm = 0x2ACFCC8
	movq	%r12, -48(%rbp)
	movl	$2, %esi
	movq	%r14, %rdi
	callq	*%rbx
	movq	-96(%rbp), %rcx
	movq	%rcx, (%r15)
	addq	$80, %rsp
	popq	%rbx
	popq	%r12
	popq	%r14
	popq	%r15
	popq	%rbp
	retq
L306:
	movabsq	$jl_undefined_var_error, %rax
	movq	%r12, %rdi
	callq	*%rax
L321:
	movabsq	$jl_undefined_var_error, %rax
	movq	%r12, %rdi
	callq	*%rax
```

but I guess it takes some time to call vecnorm from blas? which makes this faster anyway.