@KristofferC 2016-10-31
That code looks really weird.. Can you actually run `norm(a)` after that?