@KristofferC 2016-10-31
For me:

```
julia> @code_native norm(a)
	.text
Filename: tensor_ops.jl
	pushq	%rbp
	movq	%rsp, %rbp
Source line: 110
	vmovsd	(%rdi), %xmm0           # xmm0 = mem[0],zero
	vmovsd	8(%rdi), %xmm1          # xmm1 = mem[0],zero
	vmovsd	16(%rdi), %xmm2         # xmm2 = mem[0],zero
	vmulsd	%xmm0, %xmm0, %xmm0
	vxorpd	%xmm3, %xmm3, %xmm3
	vaddsd	%xmm3, %xmm0, %xmm0
	vmulsd	%xmm1, %xmm1, %xmm1
	vaddsd	%xmm1, %xmm0, %xmm0
	vmulsd	%xmm2, %xmm2, %xmm1
	vaddsd	%xmm1, %xmm0, %xmm0
Source line: 46
	vsqrtsd	%xmm0, %xmm0, %xmm0
	popq	%rbp
	retq
	nopw	%cs:(%rax,%rax)
```