@fredrikekre 2016-10-31
Wut...
now it looks like this:

```jl
julia> @code_native norm(a)
	.text
Filename: tensor_ops.jl
	pushq	%rbp
	movq	%rsp, %rbp
Source line: 110
	movsd	(%rdi), %xmm0           # xmm0 = mem[0],zero
	movsd	8(%rdi), %xmm1          # xmm1 = mem[0],zero
	movsd	16(%rdi), %xmm2         # xmm2 = mem[0],zero
	mulsd	%xmm0, %xmm0
	xorpd	%xmm3, %xmm3
	addsd	%xmm0, %xmm3
	mulsd	%xmm1, %xmm1
	addsd	%xmm3, %xmm1
	mulsd	%xmm2, %xmm2
	addsd	%xmm1, %xmm2
Source line: 46
	xorps	%xmm0, %xmm0
	sqrtsd	%xmm2, %xmm0
	popq	%rbp
	retq
	nopw	(%rax,%rax)
```