@fredrikekre 2016-10-31
Ah... I probably did `@code_native` while i had norm like:
`Base.norm(v::Vec) = sqrt(dot(a,a))` which caused an error since a isn't defined heh. Hence the force push.